Support for localIdentifier setting/capability
As described in https://docs.travis-ci.com/user/browserstack/ and
https://www.browserstack.com/automate/travisci, when using Selenium
there is a identifier associated with each local connection created,
and it is provided in the environment variable
BROWSERSTACK_LOCAL_IDENTIFIER and must be included in the settings.

The change in this commit takes the ruby selenium webdriver example
described in above links, and provides the same steps for this testem
browserstack integration.