Avoid testem browser start timeout
Testem has a default browser start timeout of 30 seconds, which hits
few times while trying to run BrowserStack tests using testem. The
default setting is being overridden with a timeout of 600 (10 minutes)
which should avoid this in most cases.

The error message that pops up when this timeout occurs is the
following:

"Browser failed to connect within 30s. testem.js not loaded?"